+++ src/Gases/s_solmod_.h
@@ -530,7 +530,7 @@ class TCGFcalc: public TSolMod
 		long int IdealProp( double *Zid );
 
         /// CGofPureGases, calculates fugacity for 1 species at (X=1)
-		long int CGcalcFugPure( double Tmin, float *Cemp, double *FugProps );  // called from DCthermo
+        long int CGcalcFugPure( double Tmin, double *Cemp, double *FugProps );  // called from DCthermo
 		long int CGFugacityPT( double *EoSparam, double *EoSparPT, double &Fugacity,
 				double &Volume, double P, double T, double &roro );
